first semi-working version of dynamic centering
This commit includes multiple changes that (just barely) get dynamic
centering to work.

NEW
- Centers multiple windows, not just a single centered window.
- this means running something like `:TreeInspect` or `:UndotreeShow`
will result in that along with the original window to be centered in
the terminal
- now triggers re-centering on WinResize instead of VimResize
- changed `toggle_centered_buffer` to `toggle_center` to reflect that
centered contents can now be more than a single window

NOTE: This is only barely a working version: it still requires some bugs
to be worked out, e.g.
- undefined behavior when you close the original window that was in
  focus when centerterm was activated
As well as some improvements, e.g.
- Specify some kind of min width (possibly for the main window) before
  centering is toggled off